requirements.txt
awscli==1.17.13
boto3==1.12.1
botocore==1.15.1
connexion==2.6.0
Flask==1.1.1
Jinja2==2.11.1
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
PyPDF2==1.26.0
PyYAML==5.2
Werkzeug==1.0.0
xmltodict==0.12.0
zappa==0.50.0
connexion[swagger-ui]
zipfile
yaml
xmltodict
